feat: re-implement delete group dialog with new file structure
