+++ src/features/groups/components/GroupCardList/GroupCard/GroupCardContent/GroupCardDropdownMenuTriggerButton/GroupCardDropdownMenuContent/DeleteGroupDialogTriggerButton/index.tsx
@@ -1,0 +1,29 @@
+import { DeleteIcon } from '@/assets/images/icons';
+import {
+  DialogRoot,
+  DialogTrigger,
+  DropdownMenuButton,
+  DropdownMenuButtonIcon,
+  DropdownMenuButtonText,
+  Icon,
+} from '@/components/ui';
+
+import React from 'react';
+
+import { DeleteGroupDialogContent } from './DeleteGroupDialogContent';
+
+export const DeleteGroupDialogTriggerButton = () => {
+  return (
+    <DialogRoot>
+      <DialogTrigger asChild>
+        <DropdownMenuButton>
+          <DropdownMenuButtonIcon>
+            <Icon icon={DeleteIcon} size={5} color="danger" />
+          </DropdownMenuButtonIcon>
+          <DropdownMenuButtonText>Delete</DropdownMenuButtonText>
+        </DropdownMenuButton>
+      </DialogTrigger>
+      <DeleteGroupDialogContent />
+    </DialogRoot>
+  );
+};
